Fix ambiguity of qFuzzyCompare on armv6
Patch from FreeBSD
https://github.com/freebsd/freebsd-ports-kde/blob/master/x11/konsole/
files/patch-src_ColorScheme.cpp
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=217079

+++ src/ColorScheme.cpp
@@ -487,7 +487,7 @@ bool ColorSchemeWallpaper::draw(QPainter &painter, const QRect rect, qreal opaci
         return false;
     }
 
-    if (qFuzzyCompare(1.0, opacity)) {
+    if (qFuzzyCompare(qreal(1.0), opacity)) {
         painter.drawTiledPixmap(rect, *_picture, rect.topLeft());
         return true;
     }
